// frontend/src/api/models/GAD7TestOutputDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GAD7TestOutputDTO
 */
export interface GAD7TestOutputDTO {
    /**
     * 
     * @type {string}
     * @memberof GAD7TestOutputDTO
     */
    testId?: string;
    /**
     * 
     * @type {string}
     * @memberof GAD7TestOutputDTO
     */
    patientId?: string;
    /**
     * 
     * @type {string}
     * @memberof GAD7TestOutputDTO
     */
    sessionId?: string;
    /**
     * 
     * @type {Date}
     * @memberof GAD7TestOutputDTO
     */
    creationDate?: Date;
    /**
     * 
     * @type {number}
     * @memberof GAD7TestOutputDTO
     */
    question1?: number;
    /**
     * 
     * @type {number}
     * @memberof GAD7TestOutputDTO
     */
    question2?: number;
    /**
     * 
     * @type {number}
     * @memberof GAD7TestOutputDTO
     */
    question3?: number;
    /**
     * 
     * @type {number}
     * @memberof GAD7TestOutputDTO
     */
    question4?: number;
    /**
     * 
     * @type {number}
     * @memberof GAD7TestOutputDTO
     */
    question5?: number;
    /**
     * 
     * @type {number}
     * @memberof GAD7TestOutputDTO
     */
    question6?: number;
    /**
     * 
     * @type {number}
     * @memberof GAD7TestOutputDTO
     */
    question7?: number;
}

/**
 * Check if a given object implements the GAD7TestOutputDTO interface.
 */
export function instanceOfGAD7TestOutputDTO(value: object): value is GAD7TestOutputDTO {
    return true;
}

export function GAD7TestOutputDTOFromJSON(json: any): GAD7TestOutputDTO {
    return GAD7TestOutputDTOFromJSONTyped(json, false);
}

export function GAD7TestOutputDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): GAD7TestOutputDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'testId': json['testId'] == null ? undefined : json['testId'],
        'patientId': json['patientId'] == null ? undefined : json['patientId'],
        'sessionId': json['sessionId'] == null ? undefined : json['sessionId'],
        'creationDate': json['creationDate'] == null ? undefined : (new Date(json['creationDate'])),
        'question1': json['question1'] == null ? undefined : json['question1'],
        'question2': json['question2'] == null ? undefined : json['question2'],
        'question3': json['question3'] == null ? undefined : json['question3'],
        'question4': json['question4'] == null ? undefined : json['question4'],
        'question5': json['question5'] == null ? undefined : json['question5'],
        'question6': json['question6'] == null ? undefined : json['question6'],
        'question7': json['question7'] == null ? undefined : json['question7'],
    };
}

export function GAD7TestOutputDTOToJSON(json: any): GAD7TestOutputDTO {
    return GAD7TestOutputDTOToJSONTyped(json, false);
}

export function GAD7TestOutputDTOToJSONTyped(value?: GAD7TestOutputDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'testId': value['testId'],
        'patientId': value['patientId'],
        'sessionId': value['sessionId'],
        'creationDate': value['creationDate'] == null ? undefined : ((value['creationDate']).toISOString()),
        'question1': value['question1'],
        'question2': value['question2'],
        'question3': value['question3'],
        'question4': value['question4'],
        'question5': value['question5'],
        'question6': value['question6'],
        'question7': value['question7'],
    };
}